[hotfix] tinyint type cast bug fix

+++ flink-formats/flink-parquet/src/main/java/org/apache/flink/formats/parquet/row/ParquetRowDataWriter.java
@@ -181,7 +181,7 @@ public class ParquetRowDataWriter {
 
         @Override
         public void write(Object value) {
-            recordConsumer.addInteger((int) value);
+            recordConsumer.addInteger((byte) value);
         }
     }
